// client/dist/widgets/common/list/src/config.ts
export enum ListLayout {
  AUTO = 'Auto',
  CUSTOM = 'Custom'
}
// ...
export enum Status {
  Default = 'DEFAULT',
  Selected = 'SELECTED',
  Hover = 'HOVER'
}